requirements.txt
datasets==2.10.1
transformers==4.26.1
seqeval==1.2.2
pyyaml>=5.4
protobuf==3.20.3
setuptools==59.5.0
pycocotools==2.0.6
pytesseract==0.3.10
chardet==5.1.0
timm==0.6.12
shapely==2.0.1
wandb>=0.13.11
opencv-python>=4.7.0.72
pytorch-lightning==1.9.4
numpy==1.24.2
tqdm>=4.65.0
torchmetrics>=0.11.3
funcy>=2.0
scikit-multilearn>=0.2.0
ultralytics==8.0.59
evaluate>=0.4.0
jiwer>=3.0.1